Groundwork: distance
We'll need to express distances (in feet only), so let's add a value object and
a convenient refinement to help us with that.

Now we can type `corridor_length = 10.feet` instead of a bland `corridor_length = 10`.

Do note that the `Distance` class is **not** very clever: the only unit that it can
handle is the imperial foot; in particular, it doesn't perform conversions to other
units (let alone units of other systems).