Groundwork: abilities
Abilities (ie. a _score_ and the _modifier_ it gives) are at the core of the game,
and characters will have at least six of them.

Abilities are value object, but they will be stored in the database, along with the
other attributes of the Character model. This is why they come with a `Ability::Type`
class, leveraging Rails 5's Attributes API.