Hatching of the `Character` class
This class doesnt't do much yet: it has a few abilities, and _partially_ implements the `#speed` method.

The implimentation is partial because a character's actual speed depends on its race, and races are not implemented yet.

Besides, `Character#speed` also depends on the object behind `Characer#armor`, but those too are still missing (hence the use of a stub in the test).