Rubocop, I hate you
I know that linters exist to hurt people's feelings, but I'm starting to
have serious doubts about the usefulness of Rubocop in general, and for
this project in particular.

Style is a (obviously) a subjective matter, and even though I can understand
how stylistic consistency can be more comfortable (and I insist on _comfortable_ –
this is not a question of objective efficiency, no matter how easily we nerds
resort to this bullshit justification for our own taste). But Ruby's delightful
expressiveness comes from its flexibility, and the way it offers several, slightly
different options for a given need, so that the best one – the most relevant
in the current context – can be used. Which means that binary rules – ”always”'s and
“never”'s – cannot easily by applied. Which is exactly what a linter does.

Besides, in a Rails app such as this one, the code base is by nature the result
of several different people's work, thanks to all the nicely generated code that
the framework provides. So, unless every developer starts adhering to the stylistic
preferences of the Rails team (if they have ones), there will inevitably be
inconsistencies, **and** an long list of files to exclude from linting in Rubocop's
configuration file.

Even though, I'm willing to keep Rubocop (for now, at least). It occasionally raises
relevant warnings, for one thing, and could prove useful should people with less
taste than me start contributing (and violating my perfectly rational and objectively
superior preferences). And, to be honest, I'm a sucker for any gamification-related
badge or good score.