Added hit points attributes to `Character`
Eventually, we'll need both the _hit point maximum_ and the _current hit points_ (since D&D characters have the bad habit of being wounded all the time).

However, the hit point maximum is not a static value: it's the (static) sum of several die rolls (one
per level, usually) and an ability modifier. Since the ability, and thus its modifier, can change, the
hit point maximum can change as well. This is why, instead of storing the hit point maximum itself, we'll
store a _base hit points_ instead; this base hit points will be the sum of the aforementioned die rolls.

As for the current hit points, we _could_ store them in the database; however, to do so, we'd need to set
actually know, when saving a new `Character`, how many hit points the character has. But since this value
usually depends on the character's hit point maximum, the data may not always be available. So, instead
of storing the current hit points, we'll store the _lost_ hit points, and differ the actual computation
of the current hit points to runtime (the current hit points, or simply _hit points_ for short, being
the hit point maximum minus the lost hit points).