**Removed** the isolation of the integration test for creating a character
Since this integration test represents the process through which a character is created, each step depends on the previous. Since theses steps are expressed as differents tests, the tests themselves depend on each other. (This is an exception to the normal rule of tests, of course.)

To ensure that all the tests share the same global state (ie. the character creation itself), we use a class variable, which is **not** reset between tests, instead of an instance variable.