`Codex` refactoring (and bug fix)
The `Codex` class used to contain a lot of logic to give it an “indifferent access” (ie. to make it so that values could be added or fetched as symbols or strings, as in a `HashWithIndifferentAccess`). This code was a bit smelly, and incomplete (we were basically on a path to re-implement `HashWithIndifferentAccess`).

To simplify all this, a subclass of `Set`, `SetWithIndifferentAccess`, has been introduced, leveraging the fact that, internally, `Set` stores its data in a `Hash` – which can easily be changed to a `HashWithIndifferentAccess.

Another change has been made to the constructor: instead of accepting a series of values, `#initialize` now requires an enum (eg. an `Array`) instead, as `Set` does. And, just like `Set`, another class method, `.[]`, is available to create a new `Codex` by passing several values.

    # Before
    Codex.new :foo, :bar:, baz
    # After
    Codex.new [:foo, :bar, :baz]
    Codex[:foo, :bar, :baz]

This changes makes the API a bit less clever, but also less surprising. It also makes it possible to create a new `Codex` with a specific `SetWithIndifferentAccess`, which may be of little value as far as the external API is concerned, but is used internally.

    Codex.new SetWithIndifferentAccess.new([:foo, :bar])
    Codex.new SetWithIndifferentAccess[:foo, :bar]

Finally, a bug has been fixed: `#expand` would crash if the `Codex` object had no groups.