Unified `Code::Pick` and `Choice`
The distinction between “a possible choice” and “an actual choice” is subtle, and lead to parallels classes. To make things simplied, those classes have been unified in a single `Pick` class.

The `Pick` represents the choice of one or several items from a given list; an extra attribute, `available`, distinguishes between a _possible_ pick (as defined in a character class, for example) and a _actual_ pick (as required made when choosing a character class)