Character creation can be suspended while proficiencies are picked
When choosing a character class during character creation (ie. when calling `CharacterCreation#choose_character_class`), it may be necessary to pick some proficiencies from a selection. For example, Barbarians can choose two proficiencies from _animal handling_, _athletics_,_intimidation_, _nature_, _perception_ and _survival_.

To do so, the choice of the character class must be suspended while the user picks the proficiencies. This is implemented by creating a _context_ for the choice of the character class itself ; this context (actually a `Fiber`) is passed to the picking of proficiencies (a `Pick` object), which _resumes_ the context once done.